pyproject.toml

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "mlx-video"
dynamic = ["version"]
description = "MLX-Video is the best package for inference and finetuning of Image-Video-Audio generation models on your Mac using MLX."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mlx>=0.22.0",
    "numpy",
    "safetensors",
    "huggingface_hub",
    "transformers[tokenizers]",
    "tqdm",
    "opencv-python>=4.12.0.88",
    "Pillow>=10.3.0",
    "mlx-vlm",
    "rich>=14.2.0",
]
license = {text="MIT"}
authors = [
    {name = "Prince Canuma", email = "prince.gdt@gmail.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/Blaizzy/mlx-video"
Repository = "https://github.com/Blaizzy/mlx-video"
Issues = "https://github.com/Blaizzy/mlx-video/issues"

[project.scripts]
"mlx_video.generate" = "mlx_video.generate:main"
"ltx_trainer.train" = "ltx_trainer.scripts.train:app"
"mlx_video.train" = "mlx_video.mlx_trainer.trainer:main"
"mlx_video.precompute" = "mlx_video.mlx_trainer.precompute:main"

[tool.setuptools.packages.find]
include = ["mlx_video*", "ltx_core*", "ltx_pipelines*", "ltx_trainer*"]

[tool.setuptools.dynamic]
version = {attr = "mlx_video.version.__version__"}

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest",
]
trainer = [
    "accelerate>=1.2.1",
    "av>=14.2.1",
    "huggingface-hub[hf-xet]>=0.31.4",
    "imageio>=2.37.0",
    "imageio-ffmpeg>=0.6.0",
    "opencv-python>=4.11.0.86",
    "optimum-quanto>=0.2.6",
    "pandas>=2.2.3",
    "peft>=0.14.0",
    "pillow-heif>=0.21.0",
    "pydantic>=2.10.4",
    "rich>=13.9.4",
    "safetensors>=0.5.0",
    "scenedetect>=0.6.5.2",
    "sentencepiece>=0.2.0",
    "torch>=2.6.0",
    "torchaudio>=2.9.0",
    "torchcodec>=0.8.1",
    "torchvision>=0.21.0",
    "typer>=0.15.1",
    "wandb>=0.19.11",
    "pyyaml>=6.0.1",
    "setuptools>=80.9.0",
]
